Oxipng optimisation for screenshots
Grabs the vanilla screenshot PNG and uses oxipng to optimise it and convert it into a progressive file. The result is losslessly smaller sized PNGs that load faster on slow networks!

Oxipng scales well with CPU cores and threads, so the more you have, the faster the optimisation process is. The default optimisation value of 2 is equal to 8 trials.