Port to Forge 41.0.99, optimise a tad, fixes
- Fix crash when opening gallery
The gallery screen seems pretty broken atm - with my simple fix it at least displays now, but it shows "Page 1 of 0" and crashes if you click the next page button.

- Fix crash when running on server
The mod's main class was referencing client-only things, potentially causing a class not found crash on dedicated servers. I've moved this to ClientModBusEvents and renamed ClientEvents to ClientForgeBusEvents

- Fixed datapack incompatibility warning

- Don't check if people are simultaneously pressing the gallery, screenshot and copy key mappings